[clients:httpAgent] Create HttpAgentSingleton
